price: int -> decimal

+++ importer/sql/warehouses/tmp_table.sql
@@ -2,7 +2,7 @@ CREATE TEMPORARY TABLE tmp_warehouses (
     product_id_1c VARCHAR(36) NOT NULL,
     stock_id_1c VARCHAR(36) NOT NULL,
     edit_date DATE,
-    price INT NOT NULL,
+    price DECIMAL(15,2) NOT NULL,
     it_rrc TINYINT(1),
     change_price_date DATE,
     load_price_date DATE,
